CMake: Fix Windows build, UnRAR dll extension
The previous UnRAR vuln fix worked for the (old, gone) Visual Studio
solution but broke the CMake build.  We didn't notice because we tested
and approved it with Visual Studio before switching to CMake.

This change switches it to use the correct macro for the
libclamunrar_iface extension which is ".dll" instead of "lib".