docker: Add script to update virus database
To reduce the need for constant database updates, also push a docker
image containing the virus database.

Ideally, this script is called every time a major virus database update
happens, but as that requires integration with other systems, lets start
by having a scheduled update.

To do this a github action workflow was added on a schedule that runs
once a month. It does require the proper secrets to be setup however and
made accessible.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>