CMake: Add CPack TGZ source package feature
Add options to build a source tarball.
This will replace Autotools' `make dist` feature.

Although CMake doesn't have an autogen step, this is still necessary if
we want to bundle in a local copy of the online documentation or if we
want to vendor dependencies from github/etc at configure time in the tarball.