Tests: Enable freshclam tests to pass as root
If you test in a docker container and don't create a user/switch to run
as something other than root, than the tests will fail.

Default freshclam behavior is to drop to run as the "clamav" user when
run as root. This change sets the DatabaseOwner to the current user, so
it the tests will pass when run as root.