RAR: limit archive extraction by dictsize
In addition to unpacked size, RAR archive file entries also have a
dictionary size. This commit checks the dictionary size when peeking at
the next file entry so that it can be skipped early if the dictionary
size is larger than our max filesize, like we do with the unpacked size.