libclamav: Only dump CRB info when using --dumpcerts
The subject and public key of each CRB signature used to be
printed in debug mode, but this will become increasingly
more noisy as we add more CRB rules. For now, given that the
number of CRB sigs is still low, only print this information
when running clamscan with the --dumpcerts parameter (or
equivalent clamd setting). If we ever get to the point where
we have hundreds of CRB rules and these log messages are
overwhelming when running with --dumpcerts, consider removing
the log messages entirely and update sigtool so that it can
nicely format and display CRB sigs with using the --decode
parameter.