FreshClam: Fix DAT file ownership if daemon-mode
The FreshClam DAT file may be owned by root if run like so:
  sudo freshclam -d

This is problematic because any future attempts to use FreshClam will
fail unless run as root/with sudo.

This commit fixes the issue by changing the file ownership if it exists
to the DatabaseOwner before dropping privileges. This has the benefit of
changing any previously existing files to be owned by the DatabaseOwner
for folks who were already affected by the bug.