Fix debug-mode assert-crash in mail parser
The mail parser uses asserts extensively to detect error conditions.
It's lazy error handling; good for prototyping but bad for production.

Release mode builds are fine in 0.103 with autotools and visual-
studio but cmake release builds will crash because asserts are enabled
even for release.

In particular this assert(0) is a possible error condition in a
malformed mail file and should be handled properly.

This resolves:
https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=31782#c2