Add issue template
The template includes a comment block at the top to direct security
issue reports towards the SECURITY.md instructions.

A comment block at the bottom provides instructions for how to share
files needed to reproduce the bug.

These comments blocks disappear when the report is submitted.

The older style markdown headers are used to match the headers printed
by the ClamConf tool, so that copy-pasted output from ClamConf looks
good in the bug report.