Test: ClamDScan ExcludePath; Valgrind
Adds a basic test to validate that ExcludePath correctly excludes a
subdirectory but does not exclude subsequent files. As with the other
ClamD/Scan tests, it will test in each mode: regular, stream, and
fdpass (if available).

Unlike the other tests, this one tests ClamDScan with Valgrind instead
of ClamD.

Refactored the clamd_test.py file to reduce duplicate code, and support
enabling and disabling valgrind when running ClamDScan and ClamD.

Add pytest to the github actions environments because the results when
using pytest are far easier to read.