Tests: fix flakey ClamD test on Windows
The non-existent file test has a hack to "expect" a wierd error message
caused by the '\v' character rather than the file not actually existing.
Recently something(?) changed and the test started reporting yet a
different message or no message.

Removing the '\v' special character fixes the test so it actually tests
a non-existent file and returns the same message as on other operating
systems.