SIS: Don't scan SIS file as new SIS file
There is loop in the SIS scanning parser where an SIS header may point
to the beginning of the file as the start of an archived file.
This would be an infinite loop if not for the scan recursion limit.

This commit fixes that by making sure that both the file records and the
individual file pointers start after the main SIS header.