libclamav, clamscan: load/unload callbacks & progress meters
Add progress callbacks to libclamav for:
- database load
- engine compile
- engine free

Add a progress bar to clamscan for load & compile.
These are disabled if you run with --debug or stdout is not a TTY or you
are using one of --quiet, --infected, or --no-summary.

Added code so you can test the engine-free callback by building with
ENABLE_ENGINE_FREE_PROGRESSBAR defined.

The compile & free progress callbacks pre-calculate the number of
tasks to complete to estimate the progress. Some tasks may take longer
than others so the progress speed my appear to vary a little.

The callbacks return type is a cl_error_t but doesn't currently do
anything. It is reserved for future use.

Minor formatting change in matcher-ac.c to counteract weird
clang-format behavior, and to make it easier to read.

Added progress callbacks and clamscan progress bars to the news.