Jenkins: create tarball in top-level pipeline
We used to generate the tarball in the build pipeline, but that required
cloning the source repo a second time in the build pipeline.

The top-level pipeline will now generate the tarball and hands it off
to downstream pipelines, that way the build pipeline will work w/out
having to know which git repository to clone.

The end goal being that our internal QA suite can trigger off of the
public git repo or an internal git repo.