tests: Fix clamd tests when path has symlink
The access-denied test and excludepath tests both relied on the full
path of the test file to be in the expected results. This fails if
you're working within a path that has a symlink because clamd and
clamdscan determine real-paths before scanning and end up sending
back the real-path in the results, not the original path.

This fixes the tests by removing the full paths from the expected
results.

I also cleaned up some type safety warnings.