Change ReceiveTimeout to use CURLOPT_LOW_SPEED_TIME
Currently ReceiveTimeout sets CURLOPT_TIMEOUT which is an absolute timeout
on the HTTP download and not particularly useful without knowing the size
of the file or the throughput available to download it.

Change it to use CURLOPT_LOW_SPEED_TIME instead, and set the related low
speed limit (CURLOPT_LOW_SPEED_LIMIT) to 1 byte per second. This will allow
the ReceiveTimeout to abort the attempt if the download is not making
any significant progress.

Restore the documentation, default and sample options back to before
2fd28e1d09f6a563564b61923e62338308e44466 and
f5d465a8644d85add515b07110c298aca10e05be.

This fixes #266 and avoids problems caused by the Ubuntu default
ReceiveTimeout of 30 seconds.