freshclam: 24-hour cooldown for 403 (forbidden)
The ClamAV CDN may send a 403 (forbidden) response for database
downloads to networks that are explicitly blocked for one reason or
another, or for download requests from out-of-date freshclam clients or
clients other than freshclam and cvdupdate.

The volume of data serving 403 responses to clients that are blocked and
retry in a tight loop is considerable.

This commit seeks to remedy that for future freshclam versions by
extending freshclam's self-regulated cool-down for 429 (retry later)
responses to include 403 (forbidden) responses.

The cooldown period if a 403 is received will be 24 hours.