Fix memory leak in zip parser error handling
The zip parser may leak a string for a zip record filename if the
the record in the central directory is the last record.

This fixes https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=31760

It also appears that there may be a related record filename leak if an
error occured when indexing the files in the central directory header,
but I don't have any test file for this but it was an obvious fix.