clamonacc: Set the names of the clamonacc-ddd and clamonacc-s(can)q(ueue) threads.
This is also not strictly necessary, but i found it rather helpful to
be able to identify and distinguish the individual clamonacc threads in
the usual Linux system tools ("ps", "top", "gdb", etc.).
Setting the thread name is already done for the threads started by the
bundled c-thread-pool library ("thread-pool-N"). Only the clamonacc
threads all identify themselfes as "clamonacc".
This commit now sets the following - shortened (due to the 16 character
limit) - thread names:
 - "clamonacc-ddd" for the ddd/inotify thread.
 - "clamonacc-sq" for the scan queue thread.
 - the main clamonacc thread currently keeps the name "clamonacc" in-
   herited from the main process.